Move `UIComponent` and related classes to its dedicated package
