+++ src/com/osudroid/multiplayer/LobbyScene.kt
@@ -2,12 +2,12 @@ package com.osudroid.multiplayer
 
 import android.net.Uri
 import android.util.Log
-import com.reco1l.andengine.*
 import com.reco1l.andengine.sprite.*
 import ru.nsu.ccfit.zuev.osu.SecurityUtils
 import com.osudroid.multiplayer.api.LobbyAPI
 import com.osudroid.multiplayer.api.RoomAPI
 import com.osudroid.utils.updateThread
+import com.reco1l.andengine.component.*
 import com.reco1l.toolkt.kotlin.async
 import org.anddev.andengine.entity.modifier.LoopEntityModifier
 import org.anddev.andengine.entity.modifier.RotationByModifier
