+++ src/com/osudroid/ui/v2/hud/editor/HUDElementOverlay.kt
@@ -11,6 +11,7 @@ import com.reco1l.andengine.text.*
 import com.reco1l.framework.ColorARGB
 import com.osudroid.ui.v2.hud.HUDElement
 import com.osudroid.utils.updateThread
+import com.reco1l.andengine.component.*
 import com.reco1l.toolkt.kotlin.*
 import org.anddev.andengine.input.touch.TouchEvent
 import ru.nsu.ccfit.zuev.osu.ResourceManager
