+++ src/com/osudroid/ui/v2/hud/elements/HUDBackButton.kt
@@ -8,6 +8,7 @@ import com.reco1l.andengine.sprite.UISprite
 import com.reco1l.framework.ColorARGB
 import com.reco1l.framework.Interpolation
 import com.osudroid.ui.v2.hud.HUDElement
+import com.reco1l.andengine.component.*
 import org.anddev.andengine.input.touch.TouchEvent
 import ru.nsu.ccfit.zuev.osu.Config
 import ru.nsu.ccfit.zuev.osu.GlobalManager
