+++ src/com/osudroid/ui/v2/modmenu/ModIcon.kt
@@ -3,6 +3,7 @@ package com.osudroid.ui.v2.modmenu
 import com.osudroid.ui.v2.*
 import com.osudroid.utils.updateThread
 import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.container.*
 import com.reco1l.andengine.shape.*
 import com.reco1l.andengine.text.*
