+++ src/com/osudroid/ui/v2/modmenu/ModMenu.kt
@@ -1,8 +1,8 @@
 package com.osudroid.ui.v2.modmenu
 
 import com.reco1l.andengine.*
-import com.reco1l.andengine.UIComponent.Companion.MatchContent
-import com.reco1l.andengine.UIComponent.Companion.FillParent
+import com.reco1l.andengine.component.UIComponent.Companion.MatchContent
+import com.reco1l.andengine.component.UIComponent.Companion.FillParent
 import com.reco1l.andengine.container.*
 import com.reco1l.andengine.shape.*
 import com.reco1l.andengine.sprite.*
