+++ src/com/osudroid/ui/v2/modmenu/ModMenuSection.kt
@@ -1,6 +1,7 @@
 package com.osudroid.ui.v2.modmenu
 
 import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.container.*
 import com.reco1l.andengine.shape.*
 import com.reco1l.andengine.text.*
