+++ src/com/osudroid/ui/v2/modmenu/ModMenuToggle.kt
@@ -2,6 +2,7 @@ package com.osudroid.ui.v2.modmenu
 
 import com.osudroid.multiplayer.*
 import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.container.*
 import com.reco1l.andengine.shape.*
 import com.reco1l.andengine.ui.*
