+++ src/com/reco1l/andengine/component/CullingMode.kt
@@ -1,4 +1,4 @@
-package com.reco1l.andengine
+package com.reco1l.andengine.component
 
 
 /**
