+++ src/com/reco1l/andengine/container/UIConstraintContainer.kt
@@ -1,6 +1,6 @@
 package com.reco1l.andengine.container
 
-import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.toolkt.kotlin.*
 import org.anddev.andengine.engine.camera.*
 import org.anddev.andengine.entity.*
