+++ src/com/reco1l/andengine/container/UIContainer.kt
@@ -1,6 +1,6 @@
 package com.reco1l.andengine.container
 
-import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import org.anddev.andengine.entity.*
 import kotlin.math.*
