+++ src/com/reco1l/andengine/container/UILinearContainer.kt
@@ -1,6 +1,6 @@
 package com.reco1l.andengine.container
 
-import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.container.Orientation.*
 import kotlin.math.*
