+++ src/com/reco1l/andengine/container/UIScrollableContainer.kt
@@ -2,6 +2,7 @@ package com.reco1l.andengine.container
 
 import androidx.annotation.*
 import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.shape.*
 import com.reco1l.framework.*
 import com.reco1l.framework.math.*
