+++ src/com/reco1l/andengine/shape/UITriangle.kt
@@ -1,7 +1,7 @@
 package com.reco1l.andengine.shape
 
-import com.reco1l.andengine.*
 import com.reco1l.andengine.buffered.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.shape.UITriangle.*
 import org.anddev.andengine.opengl.util.*
 import javax.microedition.khronos.opengles.*
