+++ src/com/reco1l/andengine/ui/CircularProgressBar.kt
@@ -1,6 +1,7 @@
 package com.reco1l.andengine.ui
 
 import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.container.*
 import com.reco1l.andengine.info.*
 import com.reco1l.framework.*
