+++ src/com/reco1l/andengine/ui/UICheckbox.kt
@@ -2,6 +2,7 @@ package com.reco1l.andengine.ui
 
 import com.edlplan.framework.easing.*
 import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.modifier.*
 import com.reco1l.andengine.shape.*
 import org.anddev.andengine.input.touch.*
