+++ src/com/reco1l/andengine/ui/UISelect.kt
@@ -2,6 +2,7 @@ package com.reco1l.andengine.ui
 
 import com.osudroid.utils.*
 import com.reco1l.andengine.*
+import com.reco1l.andengine.component.*
 import com.reco1l.andengine.container.*
 import com.reco1l.andengine.modifier.*
 import com.reco1l.andengine.shape.*
